==> Non-Automated/Code/extremes_code/step2.datacolumn_dataquality_cleanup.ipynb <==
import pandas as pd
import numpy as np
import csv
import calendar
import plotly.graph_objs as go
import plotly.express as px
from plotly.subplots import make_subplots

import dash
import dash_table
import dash_core_components as dcc
import dash_html_components as html
import datetime as dt

import os,sys,inspect
from dash.dependencies import Input, Output, State

current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

user_key=environment_variables['eshta']
pd.set_option('display.precision', 12)

data=pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats_toronto_normals_daily-1990-2020.csv')
data2= pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats-toronto-daily-1990-2020.csv')
data3= pd.read_csv(user_key['extremes_raw_unclean']+'weatherstats_toronto_normals_monthly-1990-2020.csv')


# Variables to keep 

#Keep
# date
# max_temperature
# avg_hourly_temperature
# min_temperature
# max_humidex
# min_windchill
# max_relative_humidity
# avg_hourly_relative_humidity
# min_relative_humidity
# max_wind_speed
# avg_hourly_wind_speed
# min_wind_speed
# max_pressure_sea
# avg_hourly_pressure_sea
# min_pressure_sea
# max_pressure_station
# avg_hourly_pressure_station
# min_pressure_station
# max_visibility
# avg_hourly_visibility
# min_visibility
# heatdegdays
# cooldegdays
# precipitation
# rain
# snow
# snow_on_ground

# Create new dfs for each extremes task (based on daily observed, daily normals & monthly normals)

# create new dfs
daily_normal = pd.DataFrame()
daily_observed = pd.DataFrame()
monthly_normals = pd.DataFrame()
daily_observed_mapped_version = pd.DataFrame()
daily_observed_user_defined = pd.DataFrame()

# add columns from excel daily normal to new df (daily normals)

daily_normal['date']= pd.to_datetime(data['date'])
daily_normal['max_temperature']= data['max_temperature_v']
daily_normal['min_temperature']= data['min_temperature_v']
daily_normal['max_relative_humidity']= data['max_relative_humidity_v']
daily_normal['min_relative_humidity']= data['min_relative_humidity_v']
daily_normal['max_wind_speed']= data['max_wind_speed_v']
daily_normal['min_wind_speed']= data['min_wind_speed_v']
daily_normal['precipitation']= data['precipitation_v']
daily_normal['rain']= data['rain_v']
daily_normal['snow']= data['snow_v']
daily_normal['snow_on_ground']= data['snow_on_ground_v']



# add columns from excel daily normal to new df (daily observed)

daily_observed['date']= pd.to_datetime(data2['date'])
daily_observed['avg_hourly_temperature']= data2['avg_hourly_temperature']
daily_observed['max_temperature']= data2['max_temperature']
daily_observed['min_temperature']= data2['min_temperature']
daily_observed['max_humidex']= data2['max_humidex']
daily_observed['min_windchill']= data2['min_windchill']
daily_observed['max_relative_humidity']= data2['max_relative_humidity']
daily_observed['avg_hourly_relative_humidity']= data2['avg_hourly_relative_humidity']
daily_observed['min_relative_humidity']= data2['min_relative_humidity']
daily_observed['max_wind_speed']= data2['max_wind_speed']
daily_observed['avg_hourly_wind_speed']= data2['avg_hourly_wind_speed']
daily_observed['min_wind_speed']= data2['min_wind_speed']
daily_observed['max_pressure_sea']= data2['max_pressure_sea']
daily_observed['avg_hourly_pressure_sea']= data2['avg_hourly_pressure_sea']
daily_observed['min_pressure_sea']= data2['min_pressure_sea']
daily_observed['max_pressure_station']= data2['max_pressure_station']
daily_observed['avg_hourly_pressure_station']= data2['avg_hourly_pressure_station']
daily_observed['min_pressure_station']= data2['min_pressure_station']
daily_observed['max_visibility']= data2['max_visibility']
daily_observed['avg_hourly_visibility']= data2['avg_hourly_visibility']
daily_observed['min_visibility']= data2['min_visibility']
daily_observed['cooldegdays']= data2['cooldegdays']
daily_observed['precipitation']= data2['precipitation']
daily_observed['rain']= data2['rain']
daily_observed['snow']= data2['snow']
daily_observed['snow_on_ground_v']= data2['snow_on_ground']


# add columns from excel daily normal to new df (daily observed that will be compared to daily normals)

daily_observed_mapped_version['date']= pd.to_datetime(data2['date'])
daily_observed_mapped_version['max_temperature']= data2['max_temperature']
daily_observed_mapped_version['min_temperature']= data2['min_temperature']
daily_observed_mapped_version['max_relative_humidity']= data2['max_relative_humidity']
daily_observed_mapped_version['min_relative_humidity']= data2['min_relative_humidity']
daily_observed_mapped_version['max_wind_speed']= data2['max_wind_speed']
daily_observed_mapped_version['min_wind_speed']= data2['min_wind_speed']
daily_observed_mapped_version['rain']= data2['rain']
daily_observed_mapped_version['precipitation']= data2['precipitation']
daily_observed_mapped_version['snow']= data2['snow']
daily_observed_mapped_version['snow_on_ground_v']= data2['snow_on_ground']

# add columns from excel daily normal to new df (daily observed that will be used for user-defined)

daily_observed_user_defined['date']= pd.to_datetime(data2['date'])
daily_observed_user_defined['avg_hourly_temperature']= data2['avg_hourly_temperature']
daily_observed_user_defined['max_temperature']= data2['max_temperature']
daily_observed_user_defined['min_temperature']= data2['min_temperature']
daily_observed_user_defined['precipitation']= data2['precipitation']
daily_observed_user_defined['avg_hourly_wind_speed']= data2['avg_hourly_wind_speed']
daily_observed_user_defined['avg_hourly_pressure_station']= data2['avg_hourly_pressure_station']
daily_observed_user_defined['rain']= data2['rain']
daily_observed_user_defined['snow']= data2['snow']
daily_observed_user_defined['snow_on_ground_v']= data2['snow_on_ground']
daily_observed_user_defined['max_humidex']= data2['max_humidex']
daily_observed_user_defined['min_windchill']= data2['min_windchill']
daily_observed_user_defined['avg_hourly_relative_humidity']= data2['avg_hourly_relative_humidity']
daily_observed_user_defined['avg_hourly_pressure_sea']= data2['avg_hourly_pressure_sea']


# add columns from excel daily normal to new df (monthly normals)

monthly_normals['date']= pd.to_datetime(data3['date'])
monthly_normals['max_temperature']= data3['max_temperature_v']
monthly_normals['min_temperature']= data3['min_temperature_v']
monthly_normals['max_relative_humidity']= data3['max_relative_humidity_v']
monthly_normals['min_relative_humidity']= data3['min_relative_humidity_v']
monthly_normals['max_wind_speed']= data3['max_wind_speed_v']
monthly_normals['min_wind_speed']= data3['min_wind_speed_v']
monthly_normals['precipitation']= data3['precipitation_v']
monthly_normals['rain']= data3['rain_v']
monthly_normals['snow']= data3['snow_v']
monthly_normals['snow_on_ground']= data3['snow_on_ground_v']


# Fill 0's for blank snow on ground for summer months

daily_observed['month'] = daily_observed['date'].dt.month
daily_observed_user_defined['month'] = daily_observed_user_defined['date'].dt.month
daily_observed_mapped_version['month'] = daily_observed_mapped_version['date'].dt.month


# can fill 0 for blank for snow_on_ground for all years with missing snow; 
# month must be: JUNE, JULY, AUGUST, SEPTEMBER
# need to do for 3 dfs: daily_observed; daily_observed_user_defined; daily_observed_mapped_version


daily_observed['flag_snow_null'] = np.where(daily_observed['snow_on_ground_v'].isnull(), 1,0)
daily_observed['flag_summermnth_6'] = np.where(daily_observed['month'] == 6, 1, 0) 
daily_observed['flag_summermnth_7'] = np.where(daily_observed['month'] == 7, 1, 0)
daily_observed['flag_summermnth_8'] = np.where(daily_observed['month'] == 8, 1, 0)
daily_observed['flag_summermnth_9'] = np.where(daily_observed['month'] == 9, 1, 0)
daily_observed['flag_total'] = daily_observed['flag_snow_null'] \
                                + daily_observed['flag_summermnth_6'] \
                                + daily_observed['flag_summermnth_7'] \
                                + daily_observed['flag_summermnth_8'] \
                                + daily_observed['flag_summermnth_9']

daily_observed['snow_on_ground'] = np.where(daily_observed['flag_total']==2,0,daily_observed['snow_on_ground_v'])


# can fill 0 for blank for snow_on_ground for all years with missing snow; 
# month must be: JUNE, JULY, AUGUST, SEPTEMBER
# need to do for 3 dfs: daily_observed; daily_observed_user_defined; daily_observed_mapped_version


daily_observed_user_defined['flag_snow_null'] = np.where(daily_observed_user_defined['snow_on_ground_v'].isnull(),1,0)
daily_observed_user_defined['flag_summermnth_6'] = np.where(daily_observed_user_defined['month'] == 6, 1, 0) 
daily_observed_user_defined['flag_summermnth_7'] = np.where(daily_observed_user_defined['month'] == 7, 1, 0)
daily_observed_user_defined['flag_summermnth_8'] = np.where(daily_observed_user_defined['month'] == 8, 1, 0)
daily_observed_user_defined['flag_summermnth_9'] = np.where(daily_observed_user_defined['month'] == 9, 1, 0)
daily_observed_user_defined['flag_total'] = daily_observed_user_defined['flag_snow_null'] \
                                            + daily_observed_user_defined['flag_summermnth_6'] \
                                            + daily_observed_user_defined['flag_summermnth_7'] \
                                            + daily_observed_user_defined['flag_summermnth_8'] \
                                            + daily_observed_user_defined['flag_summermnth_9']

daily_observed_user_defined['snow_on_ground'] = np.where(daily_observed_user_defined['flag_total']==2, \
                                                         0, \
                                                         daily_observed_user_defined['snow_on_ground_v'])

# can fill 0 for blank for snow_on_ground for all years with missing snow; 
# month must be: JUNE, JULY, AUGUST, SEPTEMBER
# need to do for 3 dfs: daily_observed; daily_observed_user_defined; daily_observed_mapped_version


daily_observed_mapped_version['flag_snow_null'] = np.where(daily_observed_mapped_version['snow_on_ground_v'].isnull(), 1,0)
daily_observed_mapped_version['flag_summermnth_6'] = np.where(daily_observed_mapped_version['month'] == 6, 1, 0) 
daily_observed_mapped_version['flag_summermnth_7'] = np.where(daily_observed_mapped_version['month'] == 7, 1, 0)
daily_observed_mapped_version['flag_summermnth_8'] = np.where(daily_observed_mapped_version['month'] == 8, 1, 0)
daily_observed_mapped_version['flag_summermnth_9'] = np.where(daily_observed_mapped_version['month'] == 9, 1, 0)
daily_observed_mapped_version['flag_total'] = daily_observed_mapped_version['flag_snow_null'] \
                                                + daily_observed_mapped_version['flag_summermnth_6'] \
                                                + daily_observed_mapped_version['flag_summermnth_7'] \
                                                + daily_observed_mapped_version['flag_summermnth_8'] \
                                                + daily_observed_mapped_version['flag_summermnth_9']

daily_observed_mapped_version['snow_on_ground'] = np.where(daily_observed_mapped_version['flag_total']==2, \
                                                           0, \
                                                           daily_observed['snow_on_ground_v'])


# delete unnecessary columns from all 3 dfs

del daily_observed['snow_on_ground_v']
del daily_observed['month']
del daily_observed['flag_snow_null']
del daily_observed['flag_summermnth_6']
del daily_observed['flag_summermnth_7']
del daily_observed['flag_summermnth_8']
del daily_observed['flag_summermnth_9']
del daily_observed['flag_total']

del daily_observed_user_defined['snow_on_ground_v']
del daily_observed_user_defined['month']
del daily_observed_user_defined['flag_snow_null']
del daily_observed_user_defined['flag_summermnth_6']
del daily_observed_user_defined['flag_summermnth_7']
del daily_observed_user_defined['flag_summermnth_8']
del daily_observed_user_defined['flag_summermnth_9']
del daily_observed_user_defined['flag_total']

del daily_observed_mapped_version['snow_on_ground_v']
del daily_observed_mapped_version['month']
del daily_observed_mapped_version['flag_snow_null']
del daily_observed_mapped_version['flag_summermnth_6']
del daily_observed_mapped_version['flag_summermnth_7']
del daily_observed_mapped_version['flag_summermnth_8']
del daily_observed_mapped_version['flag_summermnth_9']
del daily_observed_mapped_version['flag_total']


daily_observed

daily_observed.loc[daily_observed['date'] == '1993-06-06']


daily_observed_user_defined.loc[daily_observed_user_defined['date'] == '1993-06-06']


daily_observed_mapped_version.loc[daily_observed_mapped_version['date'] == '1993-06-06']


daily_observed.isnull().sum()

# Fill all other MISSING variables 
* snow on ground (for winter months)
* snow
* precipitation
* rain

# fill by finding the closest climate station with filled in data [for that date and variable]

# read fill in data
rainFill=pd.read_excel(user_key['extremes_raw_clean']+'DQ Reports/extremes_fillin.xlsx',sheet_name='rain')
snowFill=pd.read_excel(user_key['extremes_raw_clean']+'DQ Reports/extremes_fillin.xlsx',sheet_name='snow')
precipFill=pd.read_excel(user_key['extremes_raw_clean']+'DQ Reports/extremes_fillin.xlsx',sheet_name='precip')
snowGroundFill=pd.read_excel(user_key['extremes_raw_clean']+'DQ Reports/extremes_fillin.xlsx',sheet_name='snow_on_ground')


# loop through the data file, if the precipitation is null then 
# look through the fill in file if it has that date, if so fill in the value from fill in to the data

# make the date column the index for searching 
rainFill.set_index('date',inplace = True)
snowFill.set_index('date',inplace = True)
precipFill.set_index('Date',inplace = True)
snowGroundFill.set_index('date',inplace = True)



#PERCIPITATION 

row_iterator = daily_observed.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed.at[i,"precipitation"])):
        #print(daily_observed.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed.at[i,"precipitation"] = precipFill.loc[daily_observed.at[i,"date"],"precipitation"]
        #print(precipFill.loc[daily_observed.at[i,"date"],"precipitation"])
        
        
row_iterator = daily_observed_mapped_version.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_mapped_version.at[i,"precipitation"])):
        #print(daily_observed_mapped_version.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_mapped_version.at[i,"precipitation"] = precipFill.loc[daily_observed_mapped_version.at[i,"date"],"precipitation"]
       # print(precipFill.loc[daily_observed_mapped_version.at[i,"date"],"precipitation"])
        

row_iterator = daily_observed_user_defined.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_user_defined.at[i,"precipitation"])):
        #print(daily_observed_user_defined.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_user_defined.at[i,"precipitation"] = precipFill.loc[daily_observed_user_defined.at[i,"date"],"precipitation"]
        #print(precipFill.loc[daily_observed_user_defined.at[i,"date"],"precipitation"])
        




#RAIN 


row_iterator = daily_observed.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed.at[i,"rain"])):
        #print(daily_observed.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed.at[i,"rain"] = rainFill.loc[daily_observed.at[i,"date"],"rain"]
        #print(precipFill.loc[daily_observed.at[i,"date"],"precipitation"])
        
        
row_iterator = daily_observed_mapped_version.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_mapped_version.at[i,"rain"])):
        #print(daily_observed_mapped_version.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_mapped_version.at[i,"rain"] = rainFill.loc[daily_observed_mapped_version.at[i,"date"],"rain"]
       # print(precipFill.loc[daily_observed_mapped_version.at[i,"date"],"precipitation"])
        

row_iterator = daily_observed_user_defined.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_user_defined.at[i,"rain"])):
        #print(daily_observed_user_defined.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_user_defined.at[i,"rain"] = rainFill.loc[daily_observed_user_defined.at[i,"date"],"rain"]
        #print(precipFill.loc[daily_observed_user_defined.at[i,"date"],"precipitation"])
        



#SNOW

row_iterator = daily_observed.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed.at[i,"snow"])):
        #print(daily_observed.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed.at[i,"snow"] = snowFill.loc[daily_observed.at[i,"date"],"snow"]
        #print(precipFill.loc[daily_observed.at[i,"date"],"precipitation"])
        
        
row_iterator = daily_observed_mapped_version.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_mapped_version.at[i,"snow"])):
        #print(daily_observed_mapped_version.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_mapped_version.at[i,"snow"] = snowFill.loc[daily_observed_mapped_version.at[i,"date"],"snow"]
       # print(precipFill.loc[daily_observed_mapped_version.at[i,"date"],"precipitation"])
        

row_iterator = daily_observed_user_defined.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_user_defined.at[i,"snow"])):
        #print(daily_observed_user_defined.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_user_defined.at[i,"snow"] = snowFill.loc[daily_observed_user_defined.at[i,"date"],"snow"]
        #print(precipFill.loc[daily_observed_user_defined.at[i,"date"],"precipitation"])
        



# SNOW ON THE GROUND

row_iterator = daily_observed.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed.at[i,"snow_on_ground"])):
        #print(daily_observed.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed.at[i,"snow_on_ground"] = snowGroundFill.loc[daily_observed.at[i,"date"],"snow_on_ground"]
        #print(precipFill.loc[daily_observed.at[i,"date"],"precipitation"])
        
        
row_iterator = daily_observed_mapped_version.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_mapped_version.at[i,"snow_on_ground"])):
        #print(daily_observed_mapped_version.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_mapped_version.at[i,"snow_on_ground"] = snowGroundFill.loc[daily_observed_mapped_version.at[i,"date"],"snow_on_ground"]
       # print(precipFill.loc[daily_observed_mapped_version.at[i,"date"],"precipitation"])
        

row_iterator = daily_observed_user_defined.iterrows()  

for i, row in row_iterator:
    if (pd.isnull(daily_observed_user_defined.at[i,"snow_on_ground"])):
        #print(daily_observed_user_defined.at[i,"date"])
       # print(dataFillIn.index.values)
        #if null then replace with variable from fill in
        daily_observed_user_defined.at[i,"snow_on_ground"] = snowGroundFill.loc[daily_observed_user_defined.at[i,"date"],"snow_on_ground"]
        #print(precipFill.loc[daily_observed_user_defined.at[i,"date"],"precipitation"])
        



# Output

### For daily normals:
* Extremes raw clean

### For monthly normals:
* Extremes raw clean

### For daily observed:
* Extremes raw clean: Daily observed extravar
* Extremes raw clean: Daily observed (for normals comp)
* Extremes raw clean: Daily observed (for user defined)
* Extremes raw mapped: Daily observed (for user defined)

daily_normal.to_csv(user_key['extremes_raw_clean']+'daily_normals_cleaned.csv', index=None)
daily_observed.to_csv(user_key['extremes_raw_clean']+'daily_observed_cleaned_extravar.csv', index=None)
monthly_normals.to_csv(user_key['extremes_raw_clean']+'monthly_normals_cleaned.csv', index=None)

daily_observed_mapped_version.to_csv(user_key['extremes_raw_clean']+'daily_observed_cleaned_fornormalscomp.csv', index=None)
daily_observed_user_defined.to_csv(user_key['extremes_raw_clean']+'daily_observed_cleaned_user_defined.csv', index=None)
daily_observed_user_defined.to_csv(user_key['extremes_mapped']+'daily_observed_cleaned_user_defined.csv', index=None)
