==> Non-Automated/Code/extremes_code/step4.mapping.ipynb <==
import pandas as pd
import numpy as np
import csv
import calendar
import plotly.graph_objs as go
import plotly.express as px
from plotly.subplots import make_subplots

import dash
import dash_table
import dash_core_components as dcc
import dash_html_components as html
import datetime as dt

import os,sys,inspect
from dash.dependencies import Input, Output, State

current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

user_key=environment_variables['eshta']
pd.set_option('display.precision', 12)

data=pd.read_csv(user_key['extremes_raw_clean']+'daily_normals_cleaned.csv')
data2= pd.read_csv(user_key['extremes_raw_clean']+'daily_observed_cleaned_fornormalscomp.csv')
data_monthly= pd.read_csv(user_key['extremes_raw_clean']+'monthly_normals_cleaned.csv')
data_month = pd.read_csv(user_key['extremes_raw_clean']+'monthly_observed_cleaned_extravar.csv')

data_final = pd.DataFrame()

data_final['date']= pd.to_datetime(data['date'])

data_final['max_temperature_normal']= data['max_temperature']
data_final['max_temperature_observed']= data2['max_temperature']

data_final['min_temperature_normal']= data['min_temperature']
data_final['min_temperature_observed']= data2['min_temperature']

data_final['max_relative_humidity_normal']= data['max_relative_humidity']
data_final['max_relative_humidity_observed']= data2['max_relative_humidity']

data_final['min_relative_humidity_normal']= data['min_relative_humidity']
data_final['min_relative_humidity_observed']= data2['min_relative_humidity']

data_final['max_wind_speed_normal']= data['max_wind_speed']
data_final['max_wind_speed_observed']= data2['max_wind_speed']

data_final['min_wind_speed_normal']= data['min_wind_speed']
data_final['min_wind_speed_observed']= data2['min_wind_speed']

data_final['precipitation_normal']= data['precipitation']
data_final['precipitation_observed']= data2['precipitation']

data_final['rain_normal']= data['rain']
data_final['rain_observed']= data2['rain']

data_final['snow_normal']= data['snow']
data_final['snow_observed']= data2['snow']

data_final['snow_on_ground_normal']= data['snow_on_ground']
data_final['snow_on_ground_observed']= data2['snow_on_ground']


# map it to normals monthly 

data_Monthly_final = pd.DataFrame()

data_Monthly_final['date']= data_monthly['date']

data_Monthly_final['max_temperature_normal']= data_monthly['max_temperature']
data_Monthly_final['max_temperature_observed']= data_month['max_temperature']

data_Monthly_final['min_temperature_normal']= data_monthly['min_temperature']
data_Monthly_final['min_temperature_observed']= data_month['min_temperature']


data_Monthly_final['max_relative_humidity_normal']= data_monthly['max_relative_humidity']
data_Monthly_final['max_relative_humidity_observed']= data_month['max_relative_humidity']

data_Monthly_final['min_relative_humidity_normal']= data_monthly['min_relative_humidity']
data_Monthly_final['min_relative_humidity_observed']= data_month['min_relative_humidity']

data_Monthly_final['max_wind_speed_normal']= data_monthly['max_wind_speed']
data_Monthly_final['max_wind_speed_observed']= data_month['max_wind_speed']

data_Monthly_final['min_wind_speed_normal']= data_monthly['min_wind_speed']
data_Monthly_final['min_wind_speed_observed']= data_month['min_wind_speed']

data_Monthly_final['precipitation_normal']= data_monthly['precipitation']
data_Monthly_final['precipitation_observed']= data_month['precipitation']

data_Monthly_final['rain_normal']= data_monthly['rain']
data_Monthly_final['rain_observed']= data_month['rain']

data_Monthly_final['snow_normal']= data_monthly['snow']
data_Monthly_final['snow_observed']= data_month['snow']

data_Monthly_final['snow_on_ground_normal']= data_monthly['snow_on_ground']
data_Monthly_final['snow_on_ground_observed']= data_month['snow_on_ground']



data_Monthly_final.to_csv(user_key['extremes_mapped']+'monthly_normal_observed_mapped.csv', index=None)
data_final.to_csv(user_key['extremes_mapped']+'daily_normal_observed_mapped.csv', index=None)
